23-06-29 / DB Model Relationship & Webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 // DB Models
+import User from "../models/User";
 import Video from "../models/Video";
 
 /* Homepage - root router */
@@ -11,21 +12,32 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   // get video from DB
   const { id } = req.params;
-  const video = await Video.findById(id);
-  // return
-  if (!video)
+  const video = await Video.findById(id).populate("owner");
+  // if no video
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // Success
   return res.render("watch", { pageTitle: video.title, video });
 };
 
 /* Edit Video (GET) */
 export const getEdit = async (req, res) => {
-  // get video from DB
+  // Get video from DB
   const { id } = req.params;
   const video = await Video.findById(id);
-  // return
-  if (!video)
+  // if no video
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // Check own user
+  const {
+    user: { _id },
+  } = req.session;
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  // Success
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
@@ -36,6 +48,13 @@ export const postEdit = async (req, res) => {
   const video = await Video.exists({ _id: id });
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  // Check own user
+  const {
+    user: { _id },
+  } = req.session;
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   // Update video to DB
   const { title, description, hashtags } = req.body;
   await Video.findByIdAndUpdate(id, {
@@ -53,15 +72,23 @@ export const getUpload = (req, res) => {
 
 /* Upload Video (POST) */
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
+  // Save to DB
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       fileUrl,
       title,
       description,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -73,8 +100,24 @@ export const postUpload = async (req, res) => {
 
 /* Delete Video */
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
-  await Video.findByIdAndDelete(id);
+  const { id } = req.params; // video id
+  const {
+    user: { _id },
+  } = req.session; // user id
+  const video = await Video.findById(id);
+  // if no video
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // Check own user
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  // Success
+  await Video.findByIdAndDelete(id); // Delete video
+  const user = User.findById(_id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save(); // Delete video in user's video list
   return res.redirect("/");
 };
